### 2023-11-02 #6010 @editorialbot
[PRE REVIEW]: SCAS dashboard: A tool to intuitively and interactively analyze Slurm cluster usage
**Submitting author:** @wathom (<a href="http://orcid.org/0009-0009-3995-709X">Thomas W.</a>)
**Repository:** https://github.com/Bioinformatics-Munich/scas_dashboard
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
