### 2023-11-02 #6010 @danielskatz
👋 @wathom - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. Please feel free to use the command `@editorialbot generate pdf` after making changes to the .md file. editorialbot commands need to be the first entry in a new comment.  Once your paper compiles, please use the command `@editorialbot check references` to check the references, and if there are any errors, fix them, then use the same command again to verify that they are fixed.

See the .md file in GitHub:
