### 2023-11-02 #6010 @danielskatz
Additionally, your paper is now written as an operations paper, with this software being described as supporting operations. JOSS is for research software. I can easily imagine research applications for this software, and it would be helpful if you could add a bit of research uses of the software (in computer science, for example) to the statement of need.
