+++ joss/paper.bib
@@ -45,5 +45,6 @@ publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="44--60",
 abstract="A new cluster resource management system called Simple Linux Utility Resource Management (SLURM) is described in this paper. SLURM, initially developed for large Linux clusters at the Lawrence Livermore National Laboratory (LLNL), is a simple cluster manager that can scale to thousands of processors. SLURM is designed to be flexible and fault-tolerant and can be ported to other clusters of different size and architecture with minimal effort. We are certain that SLURM will benefit both users and system architects by providing them with a simple, robust, and highly scalable parallel job execution environment for their cluster system.",
-isbn="978-3-540-39727-4"
+isbn="978-3-540-39727-4",
+doi = {10.1007/10968987_3}
 }
