### 2023-11-03 #6017 @danielskatz
@wathom - To fix the editorialbot complaint about not finding a Statement of need section, I'm suggesting some changes to how section headings are formatted in https://github.com/Bioinformatics-Munich/scas_dashboard/pull/1 - this is what JOSS papers typically use, and I don't think it makes any difference to the formatted PDF, other than that I've promoted the conclusions header to a 1st level heading. If this seems ok, please merge it.
