### 2023-11-15 7a1d7cd
fix splitting of gres
