### 2023-11-21 #6017 @danielskatz
👋 @verolero86, if you could please create your checklist typing: `@editorialbot generate my checklist`, and check off the two items (code of conduct and COIs), I would appreciate it.
