### 2023-11-28 a963f17
Tooltip for partitions added; simplified Dockerfiles
