+++ app/00_install_packages.R
@@ -5,5 +5,5 @@
 #new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
 #if(length(new.packages)>0) install.packages(new.packages)
 install.packages(c(
-                   "rmarkdown", "shiny","shinydashboard", "dplyr", "dplyr", "readr","ggplot2","here","stringr","DT","reshape2","RColorBrewer","patchwork","httr","jsonlite","shinyjs","Rcpp"
+                   "rmarkdown", "shiny","shinydashboard", "dplyr", "readr","ggplot2","here","stringr","DT","reshape2","RColorBrewer","patchwork","httr","jsonlite","shinyjs","Rcpp","shinyBS"
 ),repos="https://packagemanager.rstudio.com/cran/__linux__/focal/2022-12-27")
