### 2023-12-05 #6017 @aturner-epcc
@wathom The link to the API documentation seems invalid. It is listed as: https://yourserver.edu/docs/ (this does not exist)
