### 2023-12-08 #6017 @wathom
> @wathom I see an failure using the installation instructions provided right at the start.
>
> > ## Build and start containers
> > Build images an run docker-compose to start the containers (nginx, frontend, backend, postgres):
> > To build: `docker compose build`
>
> ```
> ~/temp/joss/scas_dashboard/docker main                                                                                                                                                                          3.9.16 anaconda3 aturner@MBP-AT
> ❯ docker compose build
> [+] Building 1.3s (3/3) FINISHED                                                                                                                                                                                           docker:desktop-linux
>  => [scas-backend internal] load build definition from Dockerfile                                                                                                                                                                          0.0s
>  => => transferring dockerfile: 1.32kB                                                                                                                                                                                                     0.0s
>  => [scas-backend internal] load .dockerignore                                                                                                                                                                                             0.0s
>  => => transferring context: 2B                                                                                                                                                                                                            0.0s
>  => ERROR [scas-backend internal] load metadata for docker.io/rocker/shiny:4.0.5                                                                                                                                                           1.2s
> ------
>  > [scas-backend internal] load metadata for docker.io/rocker/shiny:4.0.5:
> ------
> failed to solve: rocker/shiny:4.0.5: no match for platform in manifest sha256:4e68438dc5a553b440e148ba04832007a6949361c4c9796c042599a7b2444285: not found
> ```
>
> Am I doing something wrong or missing a step?

@aturner-epcc Is this error still present, unfortunatley I can't reproduce it.
