### 2023-12-12 #6017 @aturner-epcc
I have verified that I can perform all the testing using the `export DOCKER_DEFAULT_PLATFORM=linux/amd64` option before running the `docker compose` command.
