### 2024-01-04 #6017 @verolero86
thanks for the ping @danielskatz - wrapping up EOY activities took many more cycles than expected, but now that those deadlines are behind me, I'll be able to continue the review.
