### 2024-01-07 #6017 @wathom
Dear all,
Happy new year!

> I have verified that I can perform all the testing using the `export DOCKER_DEFAULT_PLATFORM=linux/amd64` option before running the `docker compose` command. @wathom It may be worth adding a note to the README on this. Do you want to do this or do you want me to submit a PR?

Dear @aturner-epcc,
Many thanks for testing this and for providing the solution.
I am happy if you add the info and if you create a PR. Otherwise I can also add the info, just let me know.

Dear @verolero86,
Many thanks, looking forward to your comments.
