### 2024-01-30 #6017 @aturner-epcc
Sorry for the delay. PR created to address the Apple Silicon issue.

@wathom As far as I can see, the other two points I raised during my review have not been addressed. Repeating them here:

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I see mention of integration with tools like Graphana but expcted to see notes on how this compares to other commonly used tools in this area such as OpenXDMoD and Open OnDemand.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This does not seem to really be present other than "Submit an issue". What level of support can be expected (e.g. this is not funded so we will do our best in our spare time)? Are there any contribution guidelines or coding styles that need to be adhered to? What level of information is required for Issues or PRs?
