### 2024-01-30 #6017 @danielskatz
@wathom - it's generally better practice in JOSS to make changes in response to one reviewer than to wait for other reviewers. This can make the job of other reviewers easier, too.
