+++ docker/scas-frontend/print_log.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+cat $(find /var/log/shiny-server -type f -exec ls -t1 {} +| head -1)
