+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,0 +1,25 @@
+name: Bug report
+description: Report something that is broken or incorrect
+labels: bug
+body:
+  - type: textarea
+    id: description
+    attributes:
+      label: Description of the bug
+      description: A clear and concise description of what the bug is.
+    validations:
+      required: true
+  - type: textarea
+    id: logs
+    attributes:
+      label: Log files
+      description: "Please drag and drop the relevant files here.
+      Export your frontend log file with: `docker exec scas-frontend /print_log.sh`
+      Export your backend log with: `docker logs scas-backend`
+      "
+  - type: textarea
+    id: system
+    attributes:
+      label: System information
+      description: "* OS version _(e.g. Ubuntu 22.04.2 LTS)_
+        "
