### 2024-02-04 #6017 @wathom
Dear @aturner-epcc

> > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
> This does not seem to really be present other than "Submit an issue". What level of support can be expected (e.g. this is not funded so we will do our best in our spare time)? Are there any contribution guidelines or coding styles that need to be adhered to? What level of information is required for Issues or PRs?

I have added now community guidelines including Code of Conduct with the workflow how to contribute to the project, and templates for issue/bug reporting and for feature request. Please let me know if you have further suggestions on this topic.
