### 2024-02-26 #6017 @danielskatz
@aturner-epcc - there's one item on [your checklist](https://github.com/openjournals/joss-reviews/issues/6017#issuecomment-1798084928) not checked off - can you check it off now?
