### 2024-02-26 #6017 @danielskatz
👋 @verolero86 - How is your review coming? What's blocking you from checking off more items on [your checklist](https://github.com/openjournals/joss-reviews/issues/6017#issuecomment-1822835934)?
