+++ joss/paper.bib
@@ -48,3 +48,28 @@ abstract="A new cluster resource management system called Simple Linux Utility R
 isbn="978-3-540-39727-4",
 doi = {10.1007/10968987_3}
 }
+
+@ARTICLE{xdmod,
+author={Palmer, Jeffrey T. and Gallo, Steven M. and Furlani, Thomas R. and Jones, Matthew D. and DeLeon, Robert L. and White, Joseph P. and Simakov, Nikolay and Patra, Abani K. and Sperhac, Jeanette and Yearke, Thomas and Rathsam, Ryan and Innus, Martins and Cornelius, Cynthia D. and Browne, James C. and Barth, William L. and Evans, Richard T.},
+journal={Computing in Science & Engineering}, 
+title={Open XDMoD: A Tool for the Comprehensive Management of High-Performance Computing Resources}, 
+year={2015},
+volume={17},
+number={4},
+pages={52-62},
+keywords={Data warehouses;Measurement;Open source software;Open source hardware;Quality of service;Open XDMoD;XDMoD;TACC_Stats;SUPReMM;high-performance computing;HPC;HPC resource management;HPC metrics;application kernels;scientific computing},
+doi={10.1109/MCSE.2015.68}}
+
+
+@article{Hudak2018, 
+doi = {10.21105/joss.00622}, 
+url = {https://doi.org/10.21105/joss.00622}, 
+year = {2018}, 
+publisher = {The Open Journal}, 
+volume = {3}, 
+number = {25}, 
+pages = {622}, 
+author = {Dave Hudak and Doug Johnson and Alan Chalker and Jeremy Nicklas and Eric Franz and Trey Dockendorf and Brian L. McMichael}, 
+title = {Open OnDemand: A web-based client portal for HPC centers}, 
+journal = {Journal of Open Source Software} 
+}
