### 2024-02-29 #6017 @danielskatz
👋 @wscullin - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html  (Two of my three reviewers seem to have vanished, and I need at least two reviews to move this forward.)
