### 2024-03-11 #6017 @danielskatz
👋 @wscullin - if you can go ahead and generate your checklist as above, and then at least check off the COI and CoC items, I would appreciate it, in part to make sure all the permissions are working ok.
