### 2024-04-12 #6017 @danielskatz
👋 @phargogh, thanks for agreeing to review this.

Your review will be checklist based. You will have a separate checklist that you should update when carrying out your review.

As you go over the submission, please check any items that you feel have been satisfied. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if either of you require some more time. We can also use editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.Any questions/concerns please let @danielskatz know.
