### 2024-05-07 #6017 @phargogh
Thanks @danielskatz , and I'm sorry all for the delay.

@danielskatz I see that other JOSS papers have a specific _References_ section of the paper in addition to _Acknowledgements_.  This paper has an explicit _Acknowledgements_ section that includes the references.  Is this a normal format for listing references for JOSS?
