### 2024-05-07 #6017 @danielskatz
Thanks @phargogh for catching this. The paper needs to have a References header added at the end, so that the JOSS processsing that adds the actual references at the end of paper will fill them in under this header, not the Acknowledgements one.

@wathom - please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and change your paper to fix this.

> Thanks @danielskatz , and I'm sorry all for the delay.
>
> @danielskatz I see that other JOSS papers have a specific _References_ section of the paper in addition to _Acknowledgements_. This paper has an explicit _Acknowledgements_ section that includes the references. Is this a normal format for listing references for JOSS?
