### 2024-05-07 c19f65c
Update contributions.md and paper.md
