+++ joss/paper.md
@@ -223,3 +223,5 @@ We acknowledge the Institute of Computational Biology (Prof. Dr. Dr.
 Fabian Theis) at Helmholtz Munich for supporting the development of the
 software. We thank Dr. Bastian Rieck, Helmholtz Munich, for valuable
 contributions and comments to the manuscript.
+
+# References
