### 2024-05-07 #6017 @phargogh
@wathom thanks for the changes.  The proof looks complete now, with the correct references header, and I see the updates in the contribution guidelines doc now refer to github instead of gitlab.  I think it's fine to leave references to "merge requests", since folks familiar with either platform should understand well enough.

@danielskatz all boxes are now checked.  Is there anything else I need to do?
