### 2024-05-08 #6017 @danielskatz
@wathom - at this point, I'll proofread the paper and let you know the next steps. This might take me a little bit due to some other activities.
