### 2024-05-08 #6017 @danielskatz
@wathom - I'm suggesting some small changes to the paper in https://github.com/Bioinformatics-Munich/scas_dashboard/pull/3  Please merge this or let me know what you disagree with, then we can finish the process.
