### 2024-05-10 #6017 @wathom
Dear @danielskatz,

- The tagged release on Github is v1.0.0 (https://github.com/Bioinformatics-Munich/scas_dashboard/tree/v1.0.0)
- Archive on Zenodo: https://zenodo.org/records/10064783
- The Doi of the archived version is: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10064783.svg)](https://doi.org/10.5281/zenodo.10064783)
