### 2024-05-10 #6017 @danielskatz
Congratulations to @wathom (Thomas W.) and co-authors on your publication!!

And thanks to @aturner-epcc and @phargogh for reviewing this work!
JOSS depends on volunteers and wouldn't be successful without your efforts
